Move project index fs helpers into shared utils
