+++ src/plugin/src/project-index/cache.js
@@ -11,7 +11,11 @@ import {
 } from "./constants.js";
 import { defaultFsFacade } from "./fs-facade.js";
 import { createAbortGuard } from "../../../shared/abort-utils.js";
-import { isFsErrorCode, listDirectory, getFileMtime } from "./fs-utils.js";
+import {
+    isFsErrorCode,
+    listDirectory,
+    getFileMtime
+} from "../../../shared/fs-utils.js";
 
 export const PROJECT_INDEX_CACHE_SCHEMA_VERSION = 1;
 export const PROJECT_INDEX_CACHE_DIRECTORY = ".prettier-plugin-gml";
