+++ src/shared/fs-utils.js
@@ -1,0 +1,1 @@
+export * from "./utils/fs.js";
